Выполнение 2 части задания (требует доработки)
Исключение при конвертации из JSON в Java-объект.

+++ src/main/java/com/skypro/sprint1/model/PriceSum.java
@@ -4,5 +4,5 @@ import java.util.UUID;
 
 public interface PriceSum {
     UUID getProductId();
-    Integer getProductSum();
+    Long getProductSum();
 }
